fix: change to be built with tradie, added a package.json prop for setting the template (for dogfood

+++ packages/tradie/src/Template.js
@@ -35,19 +35,19 @@ export default class Template {
   }
 
   getConfig(script, args) {
-    return this.requireModule(`./config/${script}`).then(module => {
-      if (typeof module === 'function') {
-        return module(args);
+    return this.requireModule(`./lib/${script}`).then(module => {
+      if (module.__esModule) {
+        module = module.default;
       }
 
-      if (typeof module === 'object') {
-        return module;
+      if (typeof module !== 'function') {
+        throw new Error(
+          `tradie: Config "${this
+            .name}/lib/${script}" did not return a function for creating the config.`
+        );
       }
 
-      throw new Error(
-        `tradie: Config "${this
-          .name}/config/${script}" did not return a config object or function.`
-      );
+      return module(args);
     });
   }
 
@@ -61,30 +61,37 @@ Template.find = function() {
   return import(
     path.join(projectDirectory, 'package.json')
   ).then(projectMetadata => {
-    const projectDependencies = Object.keys(
-      projectMetadata.devDependencies || {}
-    );
-    const templateNames = projectDependencies.filter(name =>
-      TEMPLATE_NAME_REGEXP.test(name)
-    );
+    //this is used so we can dogfood tradie, using the previous version to run the build,
+    //rather than lerna symlinking the latest version
+    let templateName = projectMetadata['tradie-template'];
 
-    if (templateNames.length === 0) {
-      return Promise.reject(
-        // new Error( //FIXME: https://github.com/facebook/jest/issues/3699
-        `tradie: No template found in \`devDependencies\` in \`package.json\`. Please install a template e.g. \`npm install --save tradie-template-default\`.`
-        // )
+    if (!templateName) {
+      const projectDependencies = Object.keys(
+        projectMetadata.devDependencies || {}
       );
+      const templateNames = projectDependencies.filter(name =>
+        TEMPLATE_NAME_REGEXP.test(name)
+      );
+
+      if (templateNames.length === 0) {
+        return Promise.reject(
+          // new Error( //FIXME: https://github.com/facebook/jest/issues/3699
+          `tradie: No template found in \`devDependencies\` in \`package.json\`. Please install a template e.g. \`npm install --save tradie-template-default\`.`
+          // )
+        );
+      }
+
+      if (templateNames.length > 1) {
+        return Promise.reject(
+          // new Error( //FIXME: https://github.com/facebook/jest/issues/3699
+          `tradie: More than one template found in \`devDependencies\` in \`package.json\`. Please uninstall extraneous templates e.g. \`npm uninstall tradie-template-default\`.`
+          // )
+        );
+      }
+
+      templateName = templateNames[0];
     }
 
-    if (templateNames.length > 1) {
-      return Promise.reject(
-        // new Error( //FIXME: https://github.com/facebook/jest/issues/3699
-        `tradie: More than one template found in \`devDependencies\` in \`package.json\`. Please uninstall extraneous templates e.g. \`npm uninstall tradie-template-default\`.`
-        // )
-      );
-    }
-
-    const templateName = templateNames[0];
     return new Promise((resolve, reject) => {
       resolveModule(
         `${templateName}/package.json`,
